Add PetCard and ItemCardImage components
