+++ src/pages/PetsPage.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useMemo } from 'react';
+import React, { useEffect, useState, useMemo, Children } from 'react';
 import { Link } from 'react-router-dom';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
@@ -14,32 +14,9 @@ import Divider from '@mui/material/Divider';
 import Pagination from '@mui/material/Pagination';
 import { createPet, getPets } from '../api/pets';
 import { getOwners } from '../api/owners';
-import { styled } from '@mui/material/styles';
-import defaultDogPhoto from '../assets/default_dog_photo.jpg';
-import defaultCatPhoto from '../assets/default_cat_photo.jpg'
-import defaultPetPhoto from '../assets/default_pet_photo.jpg'
 import PetModalForm from '../components/PetModalForm';
 import { useApi } from '../hooks/useApi';
-
-const speciesImages = {
-  dog: defaultDogPhoto,
-  cat: defaultCatPhoto,
-  // add more species and their images here
-};
-
-const PetCard = styled(Card)(({ theme }) => ({
-  height: '100%',
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'space-between',
-  boxShadow: theme.shadows[3],
-}));
-
-const PetImage = styled('img')(({theme}) => ({
-  width: '100%',
-  objectFit: 'cover',
-  maxHeight: '200px',
-}));
+import PetCard from '../components/PetCard';
 
 
 const PetsPage = () => {
@@ -48,7 +25,7 @@ const PetsPage = () => {
   const [petsPerPage] = useState(6);
   const [openDialog, setOpenDialog] = useState(false);
   const { data: pets, loading, error } = useApi(getPets);
-  const { data: owners, loading: loadingOwners, errorPets } = useApi(getOwners);
+  const { data: owners, loading: loadingOwners, error: errorOwner } = useApi(getOwners);
 
   const handleOpenDialog = () => {
     setOpenDialog(true);
@@ -76,15 +53,14 @@ const PetsPage = () => {
 
   const paginate = (pageNumber) => setCurrentPage(pageNumber);
 
-  if (loading) return <div>Loading...</div>;
-  if (error) return <div>Error: {error.message}</div>;
+  if (loading || loadingOwners) return <div>Loading...</div>;
+  if (error || errorOwner) return <div>Error: {error.message || errorOwner.message}</div>;
 
   return (
     <Box maxWidth="lg" sx={{ py: 3 }}>
       <Typography variant="h4" gutterBottom>
         Pets
       </Typography>
-      
       <Grid container spacing={3}>
         <Grid item xs={12} sm={6} md={4} lg={3}>
           <TextField
@@ -101,61 +77,25 @@ const PetsPage = () => {
           />
         </Grid>
         <Grid item xs={2} sm={2} md={2}>
-        <Button
-          variant="contained"
-          color="primary"
-          startIcon={<AddIcon />}
-          onClick={handleOpenDialog}
-          fullWidth
-          sx={{ height: '100%' }}
-        >
-          Add Pet
-        </Button>
-      </Grid>
+          <Button
+            variant="contained"
+            color="primary"
+            startIcon={<AddIcon />}
+            onClick={handleOpenDialog}
+            fullWidth
+            sx={{ height: '100%' }}
+          >
+            Add Pet
+          </Button>
+        </Grid>
       </Grid>
       <Divider sx={{ my: 4 }} />
       <Grid container spacing={3}>
         {currentPets.map((pet) => (
           <Grid key={pet.id} item xs={12} sm={6} md={4} lg={3}>
-            <PetCard elevation={3}>
-  <PetImage src={pet.photoUrl || speciesImages[pet.species.toLowerCase()] || defaultPetPhoto} alt={pet.name} />
-  <CardContent>
-    <PetsIcon fontSize="large" color={pet.sex === 'M' ? "primary" : "secondary"} />
-    <Typography variant="h6" component="h2" gutterBottom sx={{ mt: 2 }}>
-      {pet.name}
-    </Typography>
-    { pet.owner && (
-      <Typography key={pet.owner.id} variant="body2" color="textSecondary" gutterBottom>
-        Owner: {pet.owner.name}
-      </Typography>
-    )}
-    <Typography variant="body2" color="textSecondary" gutterBottom>
-      Species: {pet.species}
-    </Typography>
-    <Typography variant="body2" color="textSecondary" gutterBottom>
-      Breed: {pet.breed}
-    </Typography>
-    <Typography variant="body2" color="textSecondary" gutterBottom>
-      Sex: {pet.sex}
-    </Typography>
-    <Typography variant="body2" color="textSecondary" gutterBottom>
-      Age: {pet.age}
-    </Typography>
-  </CardContent>
-  <Button
-    component={Link}
-    to={`/dashboard/pets/${pet.id}`}
-    variant="contained"
-    color={pet.sex === 'M' ? "primary" : "secondary"}
-    size="small"
-    fullWidth
-  >
-    View Details
-  </Button>
-</PetCard>
+            <PetCard pet={pet} elevation={3} />
           </Grid>
         ))}
-        
       </Grid>
       <Pagination
         count={Math.ceil(filteredPets.length / petsPerPage)}
